# docs/api/swagger.yaml
openapi: 3.0.0
info:
  title: Fuel Pre-Order API
  description: API for fuel pre-order and pickup mobile application
  version: 1.0.0
  contact:
    name: Fuel App Team
    email: support@fuelapp.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:3000/api
    description: Development server
  - url: https://api.fuelapp.com/api
    description: Production server
security:
  - bearerAuth: []
paths:
...
  /stations/nearby:
    get:
      tags:
        - Stations
      summary: Find nearby fuel stations
      parameters:
        - name: lat
          in: query
          required: true
          schema:
            type: number
            format: float
        - name: lng
          in: query
          required: true
          schema:
            type: number
            format: float
        - name: radius
          in: query
          schema:
            type: number
            default: 10
        - name: fuelType
          in: query
          schema:
            type: string
            enum: [gasoline, diesel, electric, lpg, cng, premium_gasoline, premium_diesel]
        - name: brand
          in: query
          schema:
            type: string
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
      responses:
        200:
          description: Nearby stations retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: object
                    properties:
                      stations:
                        type: array
                        items:
                          $ref: '#/components/schemas/Station'
                      count:
                        type: integer
  /stations/{id}:
    get:
      tags:
        - Stations
      summary: Get station details
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        200:
          description: Station details retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: object
                    properties:
                      station:
                        $ref: '#/components/schemas/Station'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Station:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        brand:
          type: string
        address:
          type: object
          properties:
            street:
              type: string
            city:
              type: string
            state:
              type: string
            country:
              type: string
            postalCode:
              type: string
            coordinates:
              type: object
              properties:
                latitude:
                  type: number
                longitude:
                  type: number
        fuelTypes:
          type: array
          items:
            type: object
            properties:
              type:
                type: string
              price:
                type: number
              currency:
                type: string
              isAvailable:
                type: boolean
        rating:
          type: object
          properties:
            average:
              type: number
            count:
              type: number
        isOpen:
          type: boolean
        distance:
          type: number
...